revise mochaPhantomJS inject and run until after callPhantom
Change operating procedure such that instead of injecting the core extensions
immediately after page load and then watching for
window.mochaPhantomJS.started in a loop (which was set by mochaPhantomJS.run())

Instead, mochaPhantomJS.run() uses window.callPhantom('mochaPhantomJS.run')
as the trigger to begin augmenting and running.

In this way, mocha is fully loaded and mochaPhantomJS is ready to be overlayed
on it, and once done we can immediately start the run.